build(deps): bump wgpu from 0.20.1 to 30.0.0

+++ src/wgpu_chunked_prefill.rs
@@ -152,7 +152,7 @@ impl WgpuChunkedProjectionPrefillPipeline {
         shape: GroupedAttentionShape,
     ) -> Result<wgpu::Buffer, WgpuChunkedProjectionPrefillError> {
         let layout = Self::layout(shape)?;
-        let maximum_bytes = u64::from(device.limits().max_storage_buffer_binding_size);
+        let maximum_bytes = device.limits().max_storage_buffer_binding_size;
         if layout.combined_bytes > maximum_bytes {
             return Err(ExternalWgpuError::DeviceBufferLimit {
                 required_bytes: layout.combined_bytes,
